fix node pool name as used by v1beta1 api
